chore: Unify doc string for references and merge all exceptions into one file
