+++ scala-native-crypto/src/java/security/AlgorithmParameters.scala
@@ -5,11 +5,8 @@ import java.util.Objects.requireNonNull
 
 abstract class AlgorithmParametersSpi {}
 
-/**
- * References:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/AlgorithmParameters.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/AlgorithmParameters
 abstract class AlgorithmParameters protected (
     spi: AlgorithmParametersSpi,
     provider: Provider,
