--- scala-native-crypto/src/java/security/InvalidKeyException.scala
@@ -1,8 +1,0 @@
-package java.security
-
-class InvalidKeyException(msg: String, cause: Throwable)
-    extends KeyException(msg, cause) {
-  def this() = this(null, null)
-  def this(msg: String) = this(msg, null)
-  def this(cause: Throwable) = this(null, cause)
-}
