--- scala-native-crypto/src/java/security/InvalidParameterException.scala
@@ -1,8 +1,0 @@
-package java.security
-
-class InvalidParameterException(msg: String, cause: Throwable)
-    extends IllegalArgumentException(msg, cause) {
-  def this() = this(null, null)
-  def this(msg: String) = this(msg, null)
-  def this(cause: Throwable) = this(null, cause)
-}
