+++ scala-native-crypto/src/java/security/Key.scala
@@ -1,10 +1,7 @@
 package java.security
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/Key.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/Key
 trait Key extends Serializable {
   def getAlgorithm(): String
