--- scala-native-crypto/src/java/security/KeyException.scala
@@ -1,8 +1,0 @@
-package java.security
-
-class KeyException(message: String, cause: Throwable)
-    extends GeneralSecurityException(message, cause) {
-  def this() = this(null, null)
-  def this(msg: String) = this(msg, null)
-  def this(cause: Throwable) = this(null, cause)
-}
