+++ scala-native-crypto/src/java/security/KeyFactory.scala
@@ -5,11 +5,8 @@ import java.util.Objects.requireNonNull
 
 abstract class KeyFactorySpi {}
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/KeyFactory.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/KeyFactory
 abstract class KeyFactory protected (
     spi: KeyFactorySpi,
     provider: Provider,
