+++ scala-native-crypto/src/java/security/KeyPair.scala
@@ -1,22 +1,11 @@
 package java.security
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/KeyPair.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/KeyPair.html
 final class KeyPair(
     publicKey: PublicKey,
     privateKey: PrivateKey
 ) extends Serializable {
   def getPrivate(): PrivateKey = privateKey
-
   def getPublic(): PublicKey = publicKey
-}
-
-object KeyPair {
-  def apply(
-      publicKey: PublicKey,
-      privateKey: PrivateKey
-  ): KeyPair = new KeyPair(publicKey, privateKey)
 }
