+++ scala-native-crypto/src/java/security/MessageDigest.scala
@@ -5,11 +5,8 @@ import java.util.Objects.requireNonNull
 
 abstract class MessageDigestSpi {}
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/MessageDigest.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/MessageDigest
 abstract class MessageDigest(
     spi: MessageDigestSpi,
     provider: Provider,
