--- scala-native-crypto/src/java/security/NoSuchProviderException.scala
@@ -1,7 +1,0 @@
-package java.security
-
-class NoSuchProviderException(msg: String, cause: Throwable)
-    extends GeneralSecurityException(msg, cause) {
-  def this() = this(null, null)
-  def this(msg: String) = this(msg, null)
-}
