+++ scala-native-crypto/src/java/security/PrivateKey.scala
@@ -2,9 +2,6 @@ package java.security
 
 import javax.security.auth.Destroyable
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/PrivateKey.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/PrivateKey
 trait PrivateKey extends AsymmetricKey with Destroyable {}
