+++ scala-native-crypto/src/java/security/Provider.scala
@@ -7,12 +7,9 @@ import java.util.{List => JList}
 import java.util.{Map => JMap}
 import java.util.{Set => JSet}
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/Provider.html
- *   - https://docs.oracle.com/en/java/javase/25/security/howtoimplaprovider.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/Provider.html
+// - https://docs.oracle.com/en/java/javase/25/security/howtoimplaprovider
 abstract class Provider(
     private val name: String,
     private val versionStr: String,
