--- scala-native-crypto/src/java/security/ProviderException.scala
@@ -1,8 +1,0 @@
-package java.security
-
-class ProviderException(msg: String, cause: Throwable)
-    extends RuntimeException(msg, cause) {
-  def this() = this(null, null)
-  def this(msg: String) = this(msg, null)
-  def this(cause: Throwable) = this(null, cause)
-}
