+++ scala-native-crypto/src/java/security/Signature.scala
@@ -8,11 +8,8 @@ import java.util.Objects.requireNonNull
 
 abstract class SignatureSpi {}
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/Signature.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/Signature.html
 abstract class Signature(
     spi: SignatureSpi,
     provider: Provider,
