--- scala-native-crypto/src/java/security/SignatureException.scala
@@ -1,8 +1,0 @@
-package java.security
-
-class SignatureException(msg: String, cause: Throwable)
-    extends GeneralSecurityException(msg, cause) {
-  def this() = this(null, null)
-  def this(msg: String) = this(msg, null)
-  def this(cause: Throwable) = this(null, cause)
-}
