+++ scala-native-crypto/src/java/security/cert/CertSelector.scala
@@ -1,10 +1,7 @@
 package java.security.cert
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/cert/CertSelector.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/cert/CertSelector
 trait CertSelector extends Cloneable {
   def `match`(cert: Certificate): Boolean
 }
