+++ scala-native-crypto/src/java/security/cert/PKIXBuilderParameters.scala
@@ -3,11 +3,8 @@ package java.security.cert
 import java.security.{InvalidParameterException, KeyStore}
 import java.util.{Set => JSet}
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/cert/PKIXBuilderParameters.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/cert/PKIXBuilderParameters
 class PKIXBuilderParameters private (
     trustAnchors: JSet[TrustAnchor],
     keyStore: KeyStore,
