+++ scala-native-crypto/src/java/security/cert/X509Certificate.scala
@@ -7,11 +7,8 @@ import java.util.{List => JList}
 import java.security.Principal
 import javax.security.auth.x500.X500Principal
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/cert/X509Certificate.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/cert/X509Certificate
 abstract class X509Certificate()
     extends Certificate("X.509")
     with X509Extension {
