+++ scala-native-crypto/src/java/security/cert/X509Extension.scala
@@ -2,11 +2,8 @@ package java.security.cert
 
 import java.util.{Set => JSet}
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/security/cert/X509Extension.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/security/cert/X509Extension
 trait X509Extension {
   def getCriticalExtensionOIDs(): JSet[String]
