--- scala-native-crypto/src/java/security/spec/InvalidParameterSpecException.scala
@@ -1,8 +1,0 @@
-package java.security.spec
-
-import java.security.GeneralSecurityException
-
-class InvalidParameterSpecException(msg: String)
-    extends GeneralSecurityException(msg, null) {
-  def this() = this(null)
-}
