+++ scala-native-crypto/src/java/security/spec/PKCS8EncodedKeySpec.scala
@@ -1,10 +1,7 @@
 package java.security.spec
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/spec/PKCS8EncodedKeySpec.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/spec/PKCS8EncodedKeySpec
 class PKCS8EncodedKeySpec(
     encodedKey: Array[Byte],
     algorithm: String
