+++ scala-native-crypto/src/java/security/spec/X509EncodedKeySpec.scala
@@ -1,12 +1,9 @@
 package java.security.spec
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/spec/X509EncodedKeySpec.html
- *
- * since JDK 9
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/java/security/spec/X509EncodedKeySpec.html
+//
+// @since JDK 9
 class X509EncodedKeySpec(
     encodedKey: Array[Byte],
     algorithm: String
