+++ scala-native-crypto/src/java/security/spec/exceptions.scala
@@ -8,3 +8,8 @@ class InvalidKeySpecException(msg: String, cause: Throwable)
   def this(msg: String) = this(msg, null)
   def this(cause: Throwable) = this(null, cause)
 }
+
+class InvalidParameterSpecException(msg: String)
+    extends GeneralSecurityException(msg, null) {
+  def this() = this(null)
+}
