+++ scala-native-crypto/src/javax/crypto/Cipher.scala
@@ -10,11 +10,8 @@ import java.security.spec.AlgorithmParameterSpec
 
 abstract class CipherSpi {}
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/javax/crypto/Cipher.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/javax/crypto/Cipher
 abstract class Cipher protected (
     spi: CipherSpi,
     provider: Provider,
