+++ scala-native-crypto/src/javax/security/auth/Destroyable.scala
@@ -1,10 +1,7 @@
 package javax.security.auth
 
-/**
- * Refs:
- *
- *   - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/javax/security/auth/Destroyable.html
- */
+// Refs:
+// - https://docs.oracle.com/en/java/javase/25/docs/api/java.base/javax/security/auth/Destroyable
 trait Destroyable {
   def destroy(): Unit
   def isDestroyed(): Boolean
