+++ scala-native-crypto/src/javax/security/auth/callback/exceptions.scala
