+++ scala-native-crypto/src/javax/security/auth/exceptions.scala
